requirements.txt
pyttsx3
vaderSentiment
textblob
flask
google-cloud-texttospeech